more resourceService methods, getAll renamed to list
