+++ internal/server/fallbackresponse.go
@@ -1,0 +1,64 @@
+package server
+
+import (
+	"fmt"
+	"io/ioutil"
+	"log"
+	"net/http"
+)
+
+// FallbackResponseWriter wraps an http.Requesthandler and surpresses
+// a 404 status code. In such case a given local file will be served.
+type FallbackResponseWriter struct {
+	WrappedResponseWriter http.ResponseWriter
+	FileNotFound          bool
+}
+
+func addFallbackHandler(handler http.HandlerFunc, fs http.FileSystem) http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		frw := FallbackResponseWriter{
+			WrappedResponseWriter: w,
+			FileNotFound:          false,
+		}
+		handler(&frw, r)
+		if frw.FileNotFound {
+			f, err := fs.Open("/index.html")
+			if err != nil {
+				log.Fatal("Failed to open index.html")
+			}
+			defer f.Close()
+			content, err := ioutil.ReadAll(f)
+			if err != nil {
+				log.Fatal("Failed to read index.html")
+			}
+
+			w.Header().Set("Content-Type", "text/html; charset=utf-8")
+			fmt.Fprint(w, string(content))
+		}
+	}
+}
+
+// Header returns the header of the wrapped response writer
+func (frw *FallbackResponseWriter) Header() http.Header {
+	return frw.WrappedResponseWriter.Header()
+}
+
+// Write sends bytes to wrapped response writer, in case of FileNotFound
+// It surpresses further writes (concealing the fact though)
+func (frw *FallbackResponseWriter) Write(b []byte) (int, error) {
+	if frw.FileNotFound {
+		return len(b), nil
+	}
+	return frw.WrappedResponseWriter.Write(b)
+}
+
+// WriteHeader sends statusCode to wrapped response writer
+func (frw *FallbackResponseWriter) WriteHeader(statusCode int) {
+
+	if statusCode == http.StatusNotFound {
+		frw.FileNotFound = true
+		return
+	}
+
+	frw.WrappedResponseWriter.WriteHeader(statusCode)
+}
