+++ internal/server/handlers.go
@@ -1,41 +1,13 @@
 package server
 
 import (
-	"fmt"
-	"io/ioutil"
-	"log"
 	"net/http"
 
 	"github.com/labstack/echo"
 	rbacv1 "k8s.io/api/rbac/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"sighupio/permission-manager/internal/kubeconfig"
 )
 
-// Header returns the header of the wrapped response writer
-func (frw *FallbackResponseWriter) Header() http.Header {
-	return frw.WrappedResponseWriter.Header()
-}
-
-// Write sends bytes to wrapped response writer, in case of FileNotFound
-// It surpresses further writes (concealing the fact though)
-func (frw *FallbackResponseWriter) Write(b []byte) (int, error) {
-	if frw.FileNotFound {
-		return len(b), nil
-	}
-	return frw.WrappedResponseWriter.Write(b)
-}
-
-// WriteHeader sends statusCode to wrapped response writer
-func (frw *FallbackResponseWriter) WriteHeader(statusCode int) {
-
-	if statusCode == http.StatusNotFound {
-		frw.FileNotFound = true
-		return
-	}
-
-	frw.WrappedResponseWriter.WriteHeader(statusCode)
-}
 
 func ListNamespaces(c echo.Context) error {
 	ac := c.(*AppContext)
@@ -44,13 +16,13 @@ func ListNamespaces(c echo.Context) error {
 		Namespaces []string `json:"namespaces"`
 	}
 
-	names, err := ac.ResourceService.NamespaceGetAll()
+	names, err := ac.ResourceService.NamespaceList()
 
 	if err != nil {
 		return err
 	}
 
-	return c.JSON(http.StatusOK, Response{
+	return ac.okResponseWithData(Response{
 		Namespaces: names,
 	})
 }
@@ -64,25 +36,25 @@ func listRbac(c echo.Context) error {
 		RoleBindings        []rbacv1.RoleBinding        `json:"roleBindings"`
 	}
 
-	clusterRoles, err := ac.Kubeclient.RbacV1().ClusterRoles().List(c.Request().Context(), metav1.ListOptions{})
+	clusterRoles, err := ac.ResourceService.ClusterRoleList()
 
 	if err != nil {
 		return err
 	}
 
-	clusterRoleBindings, err := ac.Kubeclient.RbacV1().ClusterRoleBindings().List(c.Request().Context(), metav1.ListOptions{})
+	clusterRoleBindings, err := ac.ResourceService.ClusterRoleBindingList()
 
 	if err != nil {
 		return err
 	}
 
-	roles, err := ac.Kubeclient.RbacV1().Roles("").List(c.Request().Context(), metav1.ListOptions{})
+	roles, err := ac.ResourceService.RoleList("")
 
 	if err != nil {
 		return err
 	}
 
-	roleBindings, err := ac.Kubeclient.RbacV1().RoleBindings("").List(c.Request().Context(), metav1.ListOptions{})
+	roleBindings, err := ac.ResourceService.RoleBindingList("")
 
 	if err != nil {
 		return err
@@ -125,33 +97,3 @@ func createKubeconfig(c echo.Context) error {
 	return c.JSON(http.StatusOK, Response{Ok: true, Kubeconfig: kubeCfg})
 }
 
-// FallbackResponseWriter wraps an http.Requesthandler and surpresses
-// a 404 status code. In such case a given local file will be served.
-type FallbackResponseWriter struct {
-	WrappedResponseWriter http.ResponseWriter
-	FileNotFound          bool
-}
-
-func addFallbackHandler(handler http.HandlerFunc, fs http.FileSystem) http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		frw := FallbackResponseWriter{
-			WrappedResponseWriter: w,
-			FileNotFound:          false,
-		}
-		handler(&frw, r)
-		if frw.FileNotFound {
-			f, err := fs.Open("/index.html")
-			if err != nil {
-				log.Fatal("Failed to open index.html")
-			}
-			defer f.Close()
-			content, err := ioutil.ReadAll(f)
-			if err != nil {
-				log.Fatal("Failed to read index.html")
-			}
-
-			w.Header().Set("Content-Type", "text/html; charset=utf-8")
-			fmt.Fprint(w, string(content))
-		}
-	}
-}
